# src/main/java/com/nggiabao2004/todo_websitespring/dto/AuthRequest.java
package com.nggiabao2004.todo_websitespring.dto;

import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class AuthRequest {
    @Size(min = 6, message = "Username must be at least 6 characters")
    private String username;

    @Size(min = 6, message = "Password must be at least 6 characters")
    private String password;
}