# swagger.yaml
openapi: 3.0.3
info:
  title: Todo Website API
  version: 1.0.0
  description: RESTful API for managing personal to-do tasks and user authentication.
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/auth/signup:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Username is already taken or validation error

  /api/auth/login:
    post:
      summary: Authenticate a user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: User logged in successfully (Set-Cookie header returns JSESSIONID)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User login successfully
                  username:
                    type: string
                    example: johndoe
        '401':
          description: Invalid username or password

  /api/auth/me:
    get:
      summary: Get the current authenticated user details
      tags:
        - Authentication
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved user session info
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: johndoe
        '401':
          description: Not authenticated

  /api/auth/logout:
    post:
      summary: Logout the current user
      tags:
        - Authentication
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful

  /api/todos:
    get:
      summary: Retrieve all to-do tasks for the current user
      tags:
        - Todos
      security:
        - cookieAuth: []
      responses:
        '200':
          description: An array of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized

    post:
      summary: Create a new to-do task
      tags:
        - Todos
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreateRequest'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error

  /api/todos/{id}:
    get:
      summary: Retrieve a single task by ID
      tags:
        - Todos
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Task not found or access denied

    put:
      summary: Update an existing task
      tags:
        - Todos
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Task not found or access denied

    delete:
      summary: Delete a task
      tags:
        - Todos
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found or access denied

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID

  schemas:
    AuthRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: johndoe
          minLength: 6
        password:
          type: string
          example: secret123
          minLength: 6

    TodoCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: Buy groceries
          minLength: 3
        description:
          type: string
          example: Milk, eggs, and bread

    TodoUpdateRequest:
      type: object
      required:
        - title
        - completed
      properties:
        title:
          type: string
          example: Buy groceries
          minLength: 3
        description:
          type: string
          example: Milk, eggs, and bread
        completed:
          type: boolean
          example: true

    TodoResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          example: Milk, eggs, and bread
        completed:
          type: boolean
          example: false